Move CacheEvent into its own file; move Choice impls into choice
